@@ -20,7 +20,7 @@
     "request"
   ],
   "engines": {
-    "node": ">=0.10"
+    "node": ">=4.9.1"
   },
   "main": "lib/index.js",
   "files": [
