@@ -73,7 +73,7 @@ module.exports = {
     h1.update(hmac.digest());
     h1 = h1.digest();
     var h2 = crypto.createHmac(hashAlg, secret);
-    h2.update(new Buffer(parsedSignature.params.signature, 'base64'));
+    h2.update(Buffer.from(parsedSignature.params.signature, 'base64'));
     h2 = h2.digest();
 
     /* Node 0.8 returns strings from .digest(). */
